+++ test/cases/c/fs.c
@@ -195,7 +195,7 @@ TEST_F(fs, pread_pwrite)
     ASSERT_EQ_INT(ret, 0);
 
     buf = ev_buf_make(buffer, sizeof(buffer));
-    ret = ev_file_pread(g_test_file.file, &g_test_file.token, &buf, 1, 0,
+    ret = (int)ev_file_pread(g_test_file.file, &g_test_file.token, &buf, 1, 0,
         _test_file_pread_pwrite_on_read_done);
     ASSERT_EQ_INT(ret, 0);